BTD/requirements.txt
# BTD项目依赖包

# 核心依赖
ultralytics>=8.0.0          # YOLO模型训练和推理
opencv-python>=4.5.0        # 图像处理
pyyaml>=6.0                 # YAML配置文件解析
pillow>=8.0.0               # 图像处理
numpy>=1.20.0               # 数值计算
matplotlib>=3.3.0           # 图表绘制
seaborn>=0.11.0             # 统计图表
tqdm>=4.60.0                # 进度条
pandas>=1.3.0               # 数据处理

# PyTorch (根据系统选择)
torch>=1.9.0                # PyTorch深度学习框架
torchvision>=0.10.0         # PyTorch视觉工具

# 可选依赖
psutil>=5.8.0               # 系统信息监控 (可选)
requests>=2.25.0            # HTTP请求 (用于模型下载)

# Web服务器依赖 (开发中)
# flask>=2.0.0              # Web框架
# flask-cors>=3.0.0         # 跨域支持

# 开发工具 (可选)
# pytest>=6.0.0             # 单元测试
# black>=21.0.0              # 代码格式化
# flake8>=3.8.0              # 代码检查
